source/requirements.txt
fastapi
uvicorn
pydantic
gunicorn
sqlalchemy 
sqlalchemy.orm
pymysql
requests 
python-dotenv
pytz